requirements.txt
fastapi
uvicorn
tensorflow==2.13.0
pillow
numpy
google-cloud-storage
python-multipart
gunicorn
